Fine-tuning Physics and started creating Tutorial level
Found a fix to the dash momentum bug, but still trying to fine tune the game's physics to fell more snappy and responsive Specifically the vertical acceloration feels too floaty and slow. will have to work on that more.

Also started making a tutorial level that will teach the player the basics of the game's movement system while also introducing a few level elements.