Implemented Poison
Other than working on adminstrative work and importing assets. I've been working on stuff regarding the damage system and the groundwork for the projectiles. I've been slow to making these implementations but after this. I'll merge to main, fix the conflicts and rebranch so that the rest of the team can have the bevy of micro optimizations I've made